add tests improvements
