+++ src/frontend/tests/end-to-end/curlApiGeneration.spec.ts
