+++ src/frontend/tests/end-to-end/dropdownComponent.spec.ts
@@ -19,10 +19,14 @@ test("dropDownComponent", async ({ page }) => {
     await page.waitForTimeout(5000);
     modalCount = await page.getByTestId("modal-title")?.count();
   }
-  await page.waitForTimeout(1000);
+  await page.waitForSelector('[data-testid="blank-flow"]', {
+    timeout: 30000,
+  });
 
   await page.getByTestId("blank-flow").click();
-  await page.waitForTimeout(3000);
+  await page.waitForSelector('[data-testid="extended-disclosure"]', {
+    timeout: 30000,
+  });
 
   await page.getByTestId("extended-disclosure").click();
   await page.getByPlaceholder("Search").click();
