+++ src/frontend/tests/end-to-end/langflowShortcuts.spec.ts
@@ -27,9 +27,15 @@ test("LangflowShortcuts", async ({ page }) => {
     await page.waitForTimeout(5000);
     modalCount = await page.getByTestId("modal-title")?.count();
   }
-
+  await page.waitForSelector('[data-testid="blank-flow"]', {
+    timeout: 30000,
+  });
   await page.getByTestId("blank-flow").click();
-  await page.waitForTimeout(3000);
+
+  await page.waitForSelector('[data-testid="extended-disclosure"]', {
+    timeout: 30000,
+  });
+
   await page.getByTestId("extended-disclosure").click();
   await page.getByPlaceholder("Search").click();
   await page.getByPlaceholder("Search").fill("ollama");
@@ -59,11 +65,9 @@ test("LangflowShortcuts", async ({ page }) => {
     expect(false).toBeTruthy();
   }
 
-  await page
-    .locator(
-      '//*[@id="react-flow-id"]/div[1]/div[1]/div[1]/div/div[2]/div[2]/div/div[1]/div/div[1]/div/div/div[1]',
-    )
-    .click();
+  const ollamaTitleElement = await page.getByTestId("title-Ollama").last();
+
+  await ollamaTitleElement.click();
   await page.keyboard.press("Backspace");
 
   numberOfNodes = await page.getByTestId("title-Ollama")?.count();
@@ -82,11 +86,7 @@ test("LangflowShortcuts", async ({ page }) => {
     expect(false).toBeTruthy();
   }
 
-  await page
-    .locator(
-      '//*[@id="react-flow-id"]/div[1]/div[1]/div[1]/div/div[2]/div[2]/div/div[1]/div/div[1]/div/div/div[1]',
-    )
-    .click();
+  await ollamaTitleElement.click();
   await page.keyboard.press("Backspace");
 
   await page.getByTestId("title-Ollama").click();
