+++ src/frontend/tests/end-to-end/store-shard-0.spec.ts
@@ -1,0 +1,19 @@
+import { test } from "@playwright/test";
+
+test("should exists Store", async ({ page }) => {
+  await page.goto("/");
+  await page.waitForTimeout(1000);
+
+  await page.getByTestId("button-store").isVisible();
+  await page.getByTestId("button-store").isEnabled();
+});
+
+test("should not have an API key", async ({ page }) => {
+  await page.goto("/");
+  await page.waitForTimeout(1000);
+
+  await page.getByTestId("button-store").click();
+  await page.waitForTimeout(2000);
+
+  await page.getByText("API Key Error").isVisible();
+});
