+++ src/frontend/tests/end-to-end/toggleComponent.spec.ts
@@ -19,7 +19,9 @@ test("ToggleComponent", async ({ page }) => {
     await page.waitForTimeout(5000);
     modalCount = await page.getByTestId("modal-title")?.count();
   }
-
+  await page.waitForSelector('[data-testid="blank-flow"]', {
+    timeout: 30000,
+  });
   await page.getByTestId("blank-flow").click();
 
   await page.waitForSelector('[data-testid="extended-disclosure"]', {
